+++ docs/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Next
 
+* Tensors: added Load, Save for Local tensors.
 * Skip evaluation during test of demos.
 * Fixed dogsvscat demo's inconsistent mixed datasets issue, by yielding a correct spec.